requirements.txt
langchain>=0.1.7
langchain-community>=0.0.20
langchain-openai>=0.0.5
langgraph>=0.0.17
langchain-text-splitters>=0.0.1
langchain-chroma>=0.0.1
azure-identity>=1.15.0
azure-keyvault-secrets>=4.7.0
azure-storage-blob>=12.18.3
azure-ai-ml>=1.11.1
fastapi>=0.105.0
pydantic>=2.5.3
uvicorn>=0.24.0.post1
python-dotenv>=1.0.0
chromadb>=0.4.22
tiktoken>=0.5.2
gitpython>=3.1.40
PyGithub>=2.1.1
langchainhub>=0.1.14
langchain-core>=0.1.13
numpy>=1.24.3
